backend/src/ai_engine/hibridni_model.py
"""
Ovaj modul implementira hibridni model koji kombinuje različite AI pristupe.
Možete koristiti ovaj fajl za integraciju više modela (npr. transformers, faiss, itd.)
i za eksperimentisanje s različitim arhitekturama.
"""

# Import potrebnih biblioteka
import torch
from transformers import AutoModel, AutoTokenizer
import faiss
from .model_loader import load_model

class HibridniModel:
    """
    Klasa za hibridni model koji kombinuje različite AI komponente.
    """

    def __init__(self, transformer_model_name: str, faiss_index_path: str):
        """
        Inicijalizacija hibridnog modela.

        Args:
            transformer_model_name (str): Ime pretreniranog transformers modela.
            faiss_index_path (str): Putanja do FAISS indeksa za pretragu.
        """
        # Učitavanje transformers modela i tokenizatora
        self.tokenizer = AutoTokenizer.from_pretrained(transformer_model_name)
        self.model = AutoModel.from_pretrained(transformer_model_name)

        # Učitavanje FAISS indeksa
        self.faiss_index = faiss.read_index(faiss_index_path)

    def encode_text(self, text: str):
        """
        Kodira tekst u vektorski prostor koristeći transformers model.

        Args:
            text (str): Ulazni tekst.

        Returns:
            torch.Tensor: Vektorska reprezentacija teksta.
        """
        inputs = self.tokenizer(text, return_tensors="pt")
        with torch.no_grad():
            outputs = self.model(**inputs)
        return outputs.last_hidden_state.mean(dim=1)  # Srednja vrijednost po tokenima

    def pretrazi_bazu(self, query_vector, top_k=5):
        """
        Pretražuje FAISS bazu koristeći vektorsku reprezentaciju upita.

        Args:
            query_vector (torch.Tensor): Vektorska reprezentacija upita.
            top_k (int): Broj najbližih rezultata za povratak.

        Returns:
            list: Lista ID-ova najbližih rezultata.
        """
        query_vector_np = query_vector.numpy().astype('float32')
        distances, indices = self.faiss_index.search(query_vector_np, top_k)
        return indices

    def odgovori_na_upit(self, upit: str):
        """
        Generiše odgovor na osnovu ulaznog upita koristeći hibridni pristup.

        Args:
            upit (str): Ulazni tekstualni upit.

        Returns:
            str: Generisani odgovor.
        """
        # Kodiranje upita
        query_vector = self.encode_text(upit)

        # Pretraga u FAISS bazi
        najblizi_rezultati = self.pretrazi_bazu(query_vector)

        # TODO: Kombinovati rezultate iz FAISS-a sa generativnim modelom
        odgovor = "Ovo je placeholder odgovor."
        return odgovor

    def evaluiraj_model(self, testni_podaci):
        """
        Evaluira performanse hibridnog modela na testnim podacima.

        Args:
            testni_podaci (list): Lista testnih primjera.

        Returns:
            dict: Metričke vrijednosti evaluacije (npr. tačnost, preciznost).
        """
        # TODO: Implementirati evaluaciju modela
        return {"accuracy": 0.95, "precision": 0.92, "recall": 0.90}

class HybridModel:
    def __init__(self, model_paths: dict):
        """Load multiple models from given paths"""
        self.models = {name: load_model(path) for name, path in model_paths.items()}

    def train(self, data):
        """Train all underlying models"""
        for name, model in self.models.items():
            model.fit(data)
        return True

    def predict(self, input_data):
        """Predict using all models and return combined results"""
        results = {}
        for name, model in self.models.items():
            results[name] = model.predict(input_data)
        return results

# Primjer inicijalizacije i korištenja
if __name__ == "__main__":
    # TODO: Zamijeniti sa stvarnim imenima modela i putanjama
    transformer_model = "bert-base-uncased"
    faiss_index = "path/to/faiss/index"

    hibridni_model = HibridniModel(transformer_model, faiss_index)

    upit = "Kako implementirati hibridni AI model?"
    odgovor = hibridni_model.odgovori_na_upit(upit)
    print(odgovor)