requirements.txt
# Napomena: Ovaj requirements.txt je testiran i podržan za Ubuntu/Linux okruženje.
# Neke biblioteke (npr. deepspeed) neće raditi na Windowsu bez dodatnih ručnih koraka.
# Preporučuje se koristiti Ubuntu 22.04+ i Python 3.10/3.11 za punu kompatibilnost.

# Consolidated requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
transformers==4.35.0
torch==2.7.0
python-dotenv>=0.21.0
pydantic==2.4.2
numpy==1.26.4
# Uklonjen sentence-transformers jer nije potreban za panel/AI pokretač
requests
beautifulsoup4
langdetect
gitpython
datasets
optuna
ray[tune]
prometheus-client
sqlalchemy
faiss-cpu
mlflow
deepspeed
dvc
dvc-s3
pytest
bandit
pytest-cov
pre-commit>=2.20.0
databases
asyncpg
python-jose[cryptography]
passlib[bcrypt]
psutil
GPUtil
slowapi
psycopg2-binary
python-multipart
aiosqlite
# databases[sqlite] je već naveden, ali ga eksplicitno navodim za SQLite podršku
# databases[sqlite]
